[InstCombine] change tests to show a more obvious transform possibility
The original tests were intended to show a missing transform that would
be solved by D20774:
http://reviews.llvm.org/D20774

But it's not clear that the transform for the simpler tests is a win for
all targets. Make the tests show a larger pattern that should be a win
regardless of the cost of bitcast instructions.


git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@271603 91177308-0d34-0410-b5e6-96231b3b80d8